src/shared/python-utils/monitoring/requirements.txt
# OpenTelemetry distributed tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0

# OpenTelemetry exporters
opentelemetry-exporter-jaeger-thrift==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0

# OpenTelemetry auto-instrumentation
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-psycopg2==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-kafka-python==0.43b0

# OpenTelemetry propagators
opentelemetry-propagator-b3==1.22.0

# Prometheus metrics
prometheus-client==0.19.0

# System metrics collection
psutil==5.9.8